Inherit panel and navigator regions from a base region
